checking for conflicts, moved login,signup to components folder
